// src/generated/store-information-v3/types.gen.ts
export type StoreMetafieldsGetParametersIncludeFieldsSchemaItems = 'resource_id' | 'key' | 'value' | 'namespace' | 'permission_set' | 'resource_type' | 'description' | 'owner_client_id' | 'date_created' | 'date modified';

/**
 * StoreMetafieldsGetParametersDirection
 */
export type StoreMetafieldsGetParametersDirection = 'asc' | 'desc';
// ...
export type CollectionMeta = {
    pagination?: Pagination;
    cursor_pagination?: CursorPagination;
};

/**
 * MetaFieldCollectionResponse_Batch
 *
 * Response payload for the BigCommerce API.
 *
 */
export type MetaFieldCollectionResponseBatch = {
    data?: Array<Metafield>;
    meta?: CollectionMeta;
};
// ...
export type MetafieldBasePostPutPermissionSet = 'app_only' | 'read' | 'write' | 'read_and_sf_access' | 'write_and_sf_access';

/**
 * MetafieldBase_Post_Put
 *
 * Common Metafield properties.
 *
 */
export type MetafieldBasePostPut = {
    /**
     * Determines the visibility and writeability of the field by other API consumers.
     *
     * | Value | Description |
     * | :--- | :--- |
     * | `app_only` | Private to the app that owns the field. |
     * | `read` | Visible to other API consumers. |
     * | `write` | Open for reading and writing by other API consumers. |
     * | `read_and_sf_access` | Visible to other API consumers, including on the storefront. |
     * | `write_and_sf_access` | Open for reading and writing by other API consumers, including on the storefront. |
     *
     */
    permission_set: MetafieldBasePostPutPermissionSet;
    /**
     * Namespace for the metafield, for organizational purposes.
     *
     */
    namespace: string;
    /**
     * The name of the field, for example: `location_id`, `color`.
     *
     */
    key: string;
    /**
     * The value of the field, for example: `1`, `blue`.
     *
     */
    value: string;
    /**
     * Description for the metafields.
     *
     */
    description?: string;
};
// ...
export type MetaFieldCollectionResponseBatchPostPut = {
    data?: Array<Metafield>;
    /**
     * Empty for 200 responses.
     */
    errors?: Array<unknown>;
    meta?: WriteCollectionSuccessMeta;
};
// ...
export type MetaFieldCollectionResponsePartialSuccessPostPut = {
    /**
     * Empty for 200 responses.
     */
    data?: Array<Metafield>;
    errors?: _Error;
    meta?: WriteCollectionPartialSuccessMeta;
};
// ...
export type StoreMetafieldsPutRequestBodyContentApplicationJsonSchemaItems = {
    /**
     * Determines the visibility and writeability of the field by other API consumers.
     *
     * | Value | Description |
     * | :--- | :--- |
     * | `app_only` | Private to the app that owns the field. |
     * | `read` | Visible to other API consumers. |
     * | `write` | Open for reading and writing by other API consumers. |
     * | `read_and_sf_access` | Visible to other API consumers, including on the storefront. |
     * | `write_and_sf_access` | Open for reading and writing by other API consumers, including on the storefront. |
     *
     */
    permission_set: StoreMetafieldsPutRequestBodyContentApplicationJsonSchemaItemsPermissionSet;
    /**
     * Namespace for the metafield, for organizational purposes.
     *
     */
    namespace: string;
    /**
     * The name of the field, for example: `location_id`, `color`.
     *
     */
    key: string;
    /**
     * The value of the field, for example: `1`, `blue`.
     *
     */
    value: string;
    /**
     * Description for the metafields.
     *
     */
    description?: string;
    /**
     * The ID of metafield to update.
     *
     */
    id: number;
};

/**
 * UpdateStoreMetafieldsRequestBadRequestError
 */
export type UpdateStoreMetafieldsRequestBadRequestError = {
    status?: number;
    title?: string;
    type?: string;
    detail?: string;
};

/**
 * MetaFieldCollectionDeleteResponseSuccess
 *
 * Response payload for the BigCommerce API.
 *
 */
export type MetaFieldCollectionDeleteResponseSuccess = {
    data?: Array<number>;
    /**
     * Empty for 200 responses.
     */
    errors?: Array<unknown>;
    meta?: WriteCollectionSuccessMeta;
};

/**
 * MetaFieldCollectionResponsePartialSuccess_DELETE
 *
 * Response payload for the BigCommerce API.
 *
 */
export type MetaFieldCollectionResponsePartialSuccessDelete = {
    data?: Array<number>;
    errors?: Array<_Error>;
    meta?: WriteCollectionPartialSuccessMeta;
};

/**
 * metaCollection_open
 *
 * Response metadata.
 */
export type MetaCollectionOpen = {
    [key: string]: unknown;
};

/**
 * MetafieldResponse
 *
 * Response payload for the BigCommerce API.
 *
 */
export type MetafieldResponse = {
    data?: Metafield;
    meta?: MetaCollectionOpen;
};

/**
 * NotFound
 *
 * Error payload for the BigCommerce API.
 */
export type NotFound = {
    /**
     * 404 HTTP status code.
     *
     */
    status?: number;
    /**
     * The error title describing the particular error.
     */
    title?: string;
    type?: string;
};

/**
 * MetaFieldCollectionResponse
 *
 * Response payload for the BigCommerce API.
 *
 */
export type MetaFieldCollectionResponse = {
    data?: Metafield;
    meta?: MetaCollectionOpen;
};

export type DeleteStoreMetafieldsData = {
    /**
     * Metafields ID list.
     */
    body?: Array<number>;
    path?: never;
    query?: never;
    url: '/store/metafields';
};

export type DeleteStoreMetafieldsErrors = {
    /**
     * Response object for metafields deletion with partial success.
     *
     */
    422: MetaFieldCollectionResponsePartialSuccessDelete;
};

export type DeleteStoreMetafieldsError = DeleteStoreMetafieldsErrors[keyof DeleteStoreMetafieldsErrors];

export type DeleteStoreMetafieldsResponses = {
    /**
     * Response object for metafields deletion with success.
     *
     */
    200: MetaFieldCollectionDeleteResponseSuccess;
};

export type DeleteStoreMetafieldsResponse = DeleteStoreMetafieldsResponses[keyof DeleteStoreMetafieldsResponses];

export type GetStoreMetafieldsData = {
    body?: never;
    path?: never;
    query?: {
        /**
         * Specifies the page number in a limited (paginated) list of products.
         *
         */
        page?: number;
        /**
         * Controls the number of items per page in a limited (paginated) list of products.
         *
         */
        limit?: number;
        /**
         * Filter based on a metafieldʼs key.
         */
        key?: string;
        /**
         * Filter based on comma-separated metafieldʼs keys. Could be used with vanilla `key` query parameter.
         */
        'key:in'?: Array<string>;
        /**
         * Filter based on a metafieldʼs namespaces.
         */
        namespace?: string;
        /**
         * Filter based on comma-separated metafieldʼs namespaces. Could be used with vanilla `namespace` query parameter
         */
        'namespace:in'?: Array<string>;
        /**
         * Fields to include, in a comma-separated list. The ID and the specified fields will be returned.
         */
        include_fields?: Array<StoreMetafieldsGetParametersIncludeFieldsSchemaItems>;
        /**
         * 'Query parameter that lets you filter by the minimum date modified, for example, `2024-05-14T09:34:00`, `2024-05-14T09:34:00` or `2024-05-14`. Returns metafields modified after this date.'
         *
         */
        'date_modified:min'?: string;
        /**
         * 'Query parameter that lets you filter by the maximum date modified, for example, `2024-05-14T09:34:00`, `2024-05-14T09:34:00` or `2024-05-14`. Returns metafields modified before this date.'
         *
         */
        'date_modified:max'?: string;
        /**
         * 'Query parameter that lets you filter by the minimum date created, for example, `2024-05-14T09:34:00`, `2024-05-14T09:34:00` or `2024-05-14`. Returns metafields created after this date.'
         *
         */
        'date_created:min'?: string;
        /**
         * 'Query parameter that lets you filter by the maximum date created, for example, `2024-05-14T09:34:00`, `2024-05-14T09:34:00` or `2024-05-14`. Returns metafields created before this date.'
         *
         */
        'date_created:max'?: string;
        /**
         * Sort direction. Acceptable values are: `asc`, `desc`.
         *
         */
        direction?: StoreMetafieldsGetParametersDirection;
        /**
         * A cursor indicating where to start retrieving the previous page of results. Use this parameter to paginate backward. Not required for the initial request. For subsequent requests, use the end_cursor value returned in meta.cursor_pagination from the previous response. Works with limit, direction, and other supported query parameters. When specified, offset-based pagination (page) is ignored. Cannot be used in combination with the after parameter.
         */
        before?: string;
        /**
         * A cursor indicating where to start retrieving the next page of results. Use this parameter to paginate forward. Not required for the initial request. For subsequent requests, use the start_cursor value returned in meta.cursor_pagination from the previous response. Works with limit, direction, and other supported query parameters. When specified, offset-based pagination (page) is ignored. Cannot be used in combination with the before parameter.
         */
        after?: string;
    };
    url: '/store/metafields';
};

export type GetStoreMetafieldsResponses = {
    /**
     * List of `Metafield` objects.
     *
     */
    200: MetaFieldCollectionResponseBatch;
};

export type GetStoreMetafieldsResponse = GetStoreMetafieldsResponses[keyof GetStoreMetafieldsResponses];

export type CreateStoresMetafieldsData = {
    /**
     * A `metafield` object.
     */
    body?: Array<MetafieldBasePostPut>;
    path?: never;
    query?: never;
    url: '/store/metafields';
};

export type CreateStoresMetafieldsErrors = {
    /**
     * Response object for metafields creation with partial success.
     *
     */
    422: MetaFieldCollectionResponsePartialSuccessPostPut;
};

export type CreateStoresMetafieldsError = CreateStoresMetafieldsErrors[keyof CreateStoresMetafieldsErrors];

export type CreateStoresMetafieldsResponses = {
    /**
     * List of created `Metafield` objects.
     *
     */
    200: MetaFieldCollectionResponseBatchPostPut;
};

export type CreateStoresMetafieldsResponse = CreateStoresMetafieldsResponses[keyof CreateStoresMetafieldsResponses];

export type UpdateStoreMetafieldsData = {
    body?: Array<StoreMetafieldsPutRequestBodyContentApplicationJsonSchemaItems>;
    path?: never;
    query?: never;
    url: '/store/metafields';
};

export type UpdateStoreMetafieldsErrors = {
    /**
     * Bad Request. Input is invalid.
     */
    400: UpdateStoreMetafieldsRequestBadRequestError;
    /**
     * Response object for metafields creation with partial success.
     *
     */
    422: MetaFieldCollectionResponsePartialSuccessPostPut;
};

export type UpdateStoreMetafieldsError = UpdateStoreMetafieldsErrors[keyof UpdateStoreMetafieldsErrors];

export type UpdateStoreMetafieldsResponses = {
    /**
     * Response payload for the BigCommerce API.
     */
    200: MetaFieldCollectionResponseBatchPostPut;
};

export type UpdateStoreMetafieldsResponse = UpdateStoreMetafieldsResponses[keyof UpdateStoreMetafieldsResponses];

export type DeleteStoreMetafieldData = {
    body?: never;
    headers?: {
        /**
         * The [MIME type](https://developer.mozilla.org/en-US/docs/Web/HTTP/Basics_of_HTTP/MIME_types) of the response body.
         */
        Accept?: string;
    };
    path: {
        /**
         * The ID of the `Metafield`.
         *
         */
        metafieldId: number;
    };
    query?: never;
    url: '/store/metafields/{metafieldId}';
};

export type DeleteStoreMetafieldErrors = {
    /**
     * Not found (A metafield was not found with this query).
     *
     */
    404: NotFound;
};

export type DeleteStoreMetafieldError = DeleteStoreMetafieldErrors[keyof DeleteStoreMetafieldErrors];

export type DeleteStoreMetafieldResponses = {
    /**
     * An empty response.
     *
     */
    204: {
        [key: string]: unknown;
    };
};

export type DeleteStoreMetafieldResponse = DeleteStoreMetafieldResponses[keyof DeleteStoreMetafieldResponses];

export type GetStoreMetafieldData = {
    body?: never;
    headers?: {
        /**
         * The [MIME type](https://developer.mozilla.org/en-US/docs/Web/HTTP/Basics_of_HTTP/MIME_types) of the response body.
         */
        Accept?: string;
    };
    path: {
        /**
         * The ID of the `Metafield`.
         *
         */
        metafieldId: number;
    };
    query?: never;
    url: '/store/metafields/{metafieldId}';
};

export type GetStoreMetafieldErrors = {
    /**
     * Not found (A metafield was not found with this query).
     *
     */
    404: NotFound;
};

export type GetStoreMetafieldError = GetStoreMetafieldErrors[keyof GetStoreMetafieldErrors];

export type GetStoreMetafieldResponses = {
    /**
     * A `Metafield` object.
     *
     */
    200: MetafieldResponse;
};

export type GetStoreMetafieldResponse = GetStoreMetafieldResponses[keyof GetStoreMetafieldResponses];

export type UpdateStoreMetafieldData = {
    /**
     * A `Metafield` object.
     *
     */
    body?: MetafieldBasePostPut;
    headers?: {
        /**
         * The [MIME type](https://developer.mozilla.org/en-US/docs/Web/HTTP/Basics_of_HTTP/MIME_types) of the response body.
         */
        Accept?: string;
    };
    path: {
        /**
         * The ID of the `Metafield`.
         *
         */
        metafieldId: number;
    };
    query?: never;
    url: '/store/metafields/{metafieldId}';
};

export type UpdateStoreMetafieldErrors = {
    /**
     * The resource was not found.
     *
     */
    404: NotFound;
};

export type UpdateStoreMetafieldError = UpdateStoreMetafieldErrors[keyof UpdateStoreMetafieldErrors];

export type UpdateStoreMetafieldResponses = {
    /**
     * A metafield and metadata.
     *
     */
    200: MetaFieldCollectionResponse;
};